+++ src/index.ts
@@ -1,5 +1,6 @@
 import { KNXNetConnection } from './types';
 import { KNXNetRoutingImpl } from './routing';
+import { KNXNetDiscovery } from './discovery';
 
 export function createRouting(multicastAddress?: string, port?: number): KNXNetConnection {
   const connection = new KNXNetRoutingImpl(multicastAddress, port);
@@ -16,5 +17,10 @@ export function createRouting(multicastAddress?: string, port?: number): KNXNetC
   return connection;
 }
 
-export { KNXNetConnection, KNXNetRoutingOptions } from './types';
+export function createDiscovery(): KNXNetDiscovery {
+  return new KNXNetDiscovery();
+}
+
+export { KNXNetConnection, KNXNetRoutingOptions, DiscoveryEndpoint, DiscoveryOptions } from './types';
 export { KNX_CONSTANTS } from './constants';
+export { KNXNetDiscovery } from './discovery';